requirements.txt
Authlib==1.6.3
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.1
ecdsa==0.19.1
Flask==3.1.2
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.23
python-dotenv==1.1.1
python-http-client==3.3.7
requests==2.32.5
sendgrid==6.12.4
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
